[scan] Add testPlan option to scan (#16043)
* [scan] Add testPlan option to scan

Adds a configuration option that specifies a testplan to execute Scan
with. As with test parallelisation, using this can cause a few problems
with other integrations (such as xcpretty's ability to parse the test
output). This change isn't intended as a holistic change to address
incompatible combinations of options when using the new test report
formats, just a simple option to pass in a command line option.

* Rename 'testPlan' to 'testplan'

* Limit testPlan comment to Xcode >= 11