fix(frameit) Speedup frameit by around 2x
As stated here #14626, frameit is very slow particularly because of a regression.

The involved code was responsible for creating the title/keywords of the screenshots. In summary, to create this text, this piece of code creates a very large empty image, draws the title in it, and then trims the resulting image to fit closely around the text by using `identify -format ‘%@‘`. This command is actually quite slow.

To avoid using this command, we rely on a slightly different image magick tool that lets us directly create a text image properly sized for the text. We cannot directly do that with MiniMagick high-level methods (this doesn’t seem to be supported), so we dig a bit deeper to use ImageMagick’s convert directly.

The advantage of this regression tests compared to more internal unit tests would be that they would allow heavier refactoring of the code base such as what @janpio had started here: #14458.

Fix #14626