fix(frameit) Properly lay out keywords next to titles
The previous commit broke the scenario where stack_title is false (the keyword and the title will be laid out next to each other). In that case, the baseline alignment was incorrect.

This commit deals with that specific case, and refactors the `build_text_images` method to (hopefully) make it more clear.

Accordingly, the regression tests were broken for the “keywords” case. After this fix, there is still a very small difference, but I believe it’s acceptable => I updated the fixture expectation.