[pilot] Remove redundant update_review_detail call
There are two consecutive calls to update_review_detail() inside
update_beta_app_meta(). The first call is not required, because
demo_account_required should be one of the fields for
beta_app_review_info. If demo_account_required is set in pilot,
it is merged into the beta_app_review_info hash.