Gym: fix generating macOS .app
Previously, gym would only correct detect the generated build if it was a .pkg file. If the build output a macOS .app file, gym would fail because it could not find the built product. The workaround was to set skip_package_pkg: true, but this would skip the export step, resulting in an incorrectly signed output.