Add parameter to select if compressed package should be removed
The problem is that notarize creates a zip file that is sent to the notarization service. The zipfile is then also notarized.
Then, notarize removes the zip file. If a new zipfile is created to distribute the app, it fails notarization (at least on macOS 10.15.4)
This PR provides a new configuration parameter that can be used to prevent the zipfile from being removed. This notarized zipfile can then be used for distribution.