[spaceship] Remove multi_xml dependency
This was an optional dependency for `faraday_middleware` when parsing
XML responses. As far as I can tell, and from what is being tested with
fixtures, that used to be the case but is not any longer using the App
Store Connect API.

There is a custom plist middleware, but that one is in fastlane itself
and does not make use of `multi_xml`. I believe it's used by
`#provisioning_profiles_via_xcode_api`.

https://github.com/fastlane/fastlane/blob/2.150.0.rc4/spaceship/lib/spaceship/portal/portal_client.rb#L653-L674
https://github.com/fastlane/fastlane/blob/2.150.0.rc4/spaceship/lib/spaceship/helper/plist_middleware.rb