[spaceship] Update sinatra to 2.x (uses rack 2.x)
The sinatra usage in this project is easy enough that it didn't need any
changes when updating. While being on the latest release should be
reason enough, there is one nice thing implicitly done here (not shown
since no `Gemfile.lock` is being committed), is that `rack` is now also
at `~> 2.0`, which receives updates and security fixes (although most
were backported to 1.6.x).

https://github.com/sinatra/sinatra/blob/master/CHANGELOG.md
https://github.com/rack/rack/blob/master/CHANGELOG.md

Example of a backported CVE fix.
https://nvd.nist.gov/vuln/detail/CVE-2019-16782