Clarify Sigh Arg Conflicts Error Messages
Sigh propegates an error if the user passes in two arguments that sigh
perceives as being mutually exlusive. For example, if the user passes
both an :adhoc and a :development arg, it will fail.

The error message that it currently propegates however states that "You can't enable both
:development and :adhoc".

This is problematic because the word "enable" suggests that the two
boolean arguments can't both be true when in reality sigh will still
error if given values of "false" and "false" or "false" and "true".

This commit thus removes the word "enable" and clarifies that the problem is
caused by mutually exclusive arguments.