[fastlane_core] Update the simulator device state (#16936)
* Update the simulator device state

When booting, or shutting down, we save the new state.

* Reroute errors to /dev/null

The reason why we do this is because while this instance's state may
indicate that it is Booted or Shutdown, the real state may be the
opposite. This is because the real source of truth is the device and
other code may be changing this state.

We could always check the real state with `xcrun simctl list devices`
but that takes a small bit of time and I don't think any real problem
will occur if we ignore the error.