Modify `Net::HTTPGenericRequest` monkeypatch so that only Apple domains are affected
When using fastlane in gem form inside another project, this monkeypatch is very broad, and can cause existing code to break that relies on `Net::HTTP`'s existing header behaviour.
This modification ensures that the skipping of the Content-Type header only occurs for Apple subdomains, which covers the patch's intended targets.