Use UI.error instead of puts to report config item validation exception
This will enable testing that the user sees the error message. Besides,
it's a better experience: Fastlane will color the output making it
clearer that there's been an error.

Decided to use error instead of user_error! because this is part of an
interactive process that should continue after the error is shown and
useuser_error! would make Fastlane exit.