[fastlane_core] improve config item validation conversion and use UI.error instead of puts to report validation exception (#18088)
* Use UI.error instead of puts to report config item validation exception

This will enable testing that the user sees the error message. Besides,
it's a better experience: Fastlane will color the output making it
clearer that there's been an error.

Decided to use error instead of user_error! because this is part of an
interactive process that should continue after the error is shown and
useuser_error! would make Fastlane exit.

* Add failing tests to reproduce configuration item issue

* Auto-convert config value received by the interactive prompt

* Updated tests for failed input

Co-authored-by: Josh Holtz <josh@rokkincat.com>