Resolve issue with current team_id being incorrectly reported
Previously, the team_id method used `teams[0]['contentProvider']['contentProviderId']` to determine the team ID that is in use.
`teams` ultimately derives from the `user_details_data` session information, but critically, it is manually sorted by name before being returned.
This means that the first element in `teams` does not have a semantic meaning, and may only by chance reflect the currently selected team.

This fix uses the `user_details_data['sessionToken']['contentProviderId']` directly to determine the current team, as this attribute always reflects the ID of the current team.
The `user_details_data` is typically cached for performance, but as this data changes when switching team.
To prevent potential confusion and stale info this cache is cleared on manual team switch with `team_id=`.

Specs have been updated to test this usecase.