[fastlane_core][pilot][deliver] Improved transporter logging to pilot (#12259)
* useful failure message

This allows us to get the transporter error message in UI.user_error in pilot

* add note

* remove multiple returns, add method to get errors

Went back to the original functionality with the returns. Added new methods that let us get the transporter errors

* map errors to displayable_errors

This gets rid of errors_to_show but we are still getting errors that are essentially double printed.
What's happening is the error line(s) from transporter are getting put into @errors, and then UI.crash! is being rescued and and that is being put into @errors as well.

On top of this there are a few spots that either check if @errors is > 0 and prints the errors/tells you there were errors and also places that check the exit status and then prints the errors/tells you there were errors

* include build and version number with duplicate binary message

Helps the user with the best course of action in order to resolve duplicate binary message from transporter.

* Fix lint errors

* remove uneeded @errors_to_show

looks like I forgot to remove this when I made the change not to use it anymore

* Remove elsif block

remove elsif check for redundant binary to put in different PR

* close if

* Also updating for deliver

Co-authored-by: Jan Piotrowski <piotrowski+git@gmail.com>
Co-authored-by: Josh Holtz <josh@rokkincat.com>