Update update_plist.rb
In the example your are using a block argument to update the values of keys in an Info.plist file. 
The original code was using symbol keys (e.g., `:BRAZE_API_KEY_IOS`) to update these values, which resulted in the new values being added to the existing array for each key, instead of replacing the value with a new string.

To fix this issue, we have changed the symbol keys to string keys (e.g., `'BRAZE_API_KEY_IOS'`), which allows us to directly assign a new string value to each key using the `=` operator. This will replace the existing value for each key with the new string value, instead of appending the new value to an existing array.

By using string keys instead of symbol keys, we can ensure that the values in the Info.plist file are updated correctly and that we avoid any unintended side effects from modifying arrays instead of strings.