fix the loading of certain passwords from the keychain (fixes #21817)
If a password contains certain characters (such as a backslash), the
security command line tool will output the password in hexadecimal
format followed by a quoted escaped string.

Version 0.1.5 of the security gem supports this format, so updating to
this version will make it so fastlane can properly load saved passwords
that contain these special characters.