Update simctl command in device_manager.rb
When run_tests executes, it attempts to get a list of installed runtimes in json format. This fails with Xcode 14.n and 15.n.
The reason is that the following command is used to get the list of runtimes:
`xcrun simctl list runtimes -j`.
I'm unsure if this worked in older versions of Xcode, but it no longer does. The output modifier must come BEFORE `runtimes` in the command, e.g. `xcrun simctl list -j runtimes`