Fix bug detecting configurations from workspace
When enumerating the available configurations starting from an xcworkspace, we
have to interpret the xcodeproj file references relative to the path of the
xcworkspace. The code already handles this properly in the `project_paths`
method, but for some reason doesn't qualify the paths correctly in the
`configurations` method.

This has likely gone silently undetected because the bug only manifests itself
when the xcworkspace is in a different directory from fastlane's CWD.