Decouple import_from_git tests from each other
Before these changes, the same temp repo was being used and modified by
the tests, so they had to be executed in the same order every time.

Since recreating the temp repo takes quite a while, this created it only
once, and then creates a copy for each test execution.