[trainer] Support parsing test results from SwiftTesting (#29463)
* Extract parsing methods into dedicated files

Extract parsing of PLIST and Parsing of --legacy xcresulttool output each in their own dedicated file, on preparation for refactoring and introducing new xcresult parsing method for Xcode 16

* Add new XCResult parser + tests

* Clean large attachments and log data from xcresult

To avoid storing useless large data from fixture in git, while it's not used for our test cases

* Fix extracting data when argument child nodes

* Use different <testcase name=…> if using argument

To ensure that tools that parse the generated JUnit would consider those as different test cases, instead of multiple runs of the same test case.

We already have <property name="argument" value="…"> to help us distinguish, but in practice most of the tools that read JUnit XML files will now know about that property or account for it, and only look at the `<testcase classname=…name=…>` attributes to differentiate cases.

* Extract each Trainer::XCResult::* nested class

Into a dedicated file, for easier browsing of the code and smaller files to scroll through.

* Fix JUnitGenerator test

* Move `supports_xcode16_xcresulttool?` to Helper

* Fix `rubocop` violations

* Exclude tests requiring xcresulttool from windows

* Remove unused require

* Skip tests for new XCResult if CI don't support it

* Better report failed xcresulttool calls

* Make tests pass with earlier Xcode

So that even you run the tests on CI with older versions of Xcode (that don't need nor know the `--legacy` flag) would still run and pass those tests

* Improve shell command call

Use Open3 to run the command instead of backticks
 - So that we can capture the error if it happens
 - So that the code is more resilient against special characters
 - So that we avoid the involvement of an intermediate shell (that would have to interpret the string into parameters before passing it to the process, requiring us to `Shellwords.escape` it in the first place… instead of just passing the parameters to the process directly without involving some unnecessary shell escaping + parsing intermediary)

* Commit dummy project used to generate fixture

Even if it's not directly used/referenced as a fixture by the specs themselves, it can still be useful to have it around to make it easier to add more tests in the project then regenerate the `.xcresult` fixture with more test results cases

* Add test containing tricky characters

Both in its name (emojis) and parameters (ASCII characters 0–128)

* Update .xcresult fixture with new data

* ignore `.sqlite3` files from `.xcresult` fixtures

* Escape special chars when generating JUnit XML

* Regenerate JUnit fixtures to match new xcresult

* Update test: add error message with special chars

To ensure that our sanitizing is not just working for test parameters, but also for error messages and similar strings that could be encountered elsewhere in the generated XML.

This not only include the special, non-printable characters with unexpected ASCII values in the custom error message, but also ensure we include `>` and `<=` characters (which will have to be XML-escaped by REXML to be rendered properly)

This also incidentally validates that a `#expect(…)` with a custom error message is properly taken into account when as a custom error message when parsing the `.xcresult` and generating the JUnit XML

* Fix code comment location

* Only commit database.sqlite3 file for new fixture

The `*.xcresult/database.sqlite3` database file contains the same data as what's serialized in the `.xcresult/Data/*` files, and is just an alternate (and more modern and disk-space-optimized) representation of the same info that the latest `xcresulttool` supports

So we might as well commit just the database instead of the whole Data/* folder and its files, not only to make it take less space on disk, but also to make the git diff and logs easier to deal with.

* Attach <properties> to each <testsuite> instead

As apparently JUnit conventions don't allow the root <testsuites> element to contain <properties> elements, only <testsuite> child elements can

h/t @TheMetalCode for noticing in https://github.com/fastlane/fastlane/pull/29463#discussion_r1983844584

* Fix documentation style violation

`trainer` -> _trainer_

* Fix rubocop violations

* Update doc comments in .gitignore

* Update doc comment in legacy_xcresult.rb

Co-authored-by: Jason Hagglund <jhagglund@doximity.com>

* Add some mode code comments

Especially to be clearer about the Repetition type

* Rename variable to nicer name

As suggested in https://github.com/fastlane/fastlane/pull/29463/files#r1990796517

---------

Co-authored-by: Jason Hagglund <jhagglund@doximity.com>