Add keychain_name param to setup_ci action (#29482)
* Add keychain_name param to setup_ci action

Before this commit it was impossible to provide a `name` param to the `create_keychain` action which is called by the `setup_ci` action. This commit adds a `keychain_name` param to the `setup_ci` action which will be passed as `name` to the `create_keychain` action.

* Do required setup by moving test

It seems like some more setup was needed for the tests to pass on the CI. By moving the test, the setup should already be done.

* Trim whitespace