UI.message number of objects being downloaded
For S3 buckets that contain a lot of objects (e.g. a single bucket used for many apps in the same teamID, with many certs and profiles to download), the `download` step can take a bit of time.
Having no output at all after the Summary table can easily make it confusing as to why _fastlane match_ seems to be blocked. Adding this log will at least help explain what's happening and why the log output seems blocked for a time, making it less confusing/alarming.