[scan] Fix build_for_testing failure to fail
Scan's `ErrorHandler.handle_build_error` has explicit logic that
checks for `/Executed/` pattern in logs. It seems to be meant for case
where tests will be run with same command and so we would want to wait
for test results parsing instead. But this is causing scan to not to
fail under `build_for_testing` configuration when there is such
pattern in logs (eg. some file name contains it).

In order to fix this, I propose to explicitly check for exit status
code in `build_for_testing` mode and produce an error if status is not
ok.

Fixes #20685